+++ worker/src/__tests__/evalService.filtering.test.ts
@@ -15,7 +15,7 @@ import { randomUUID } from "crypto";
 import Decimal from "decimal.js";
 import { afterAll, test as baseTest, beforeAll, describe } from "vitest";
 import { z } from "zod";
-import { createEvalJobs } from "../ee/evaluation/evalService";
+import { createEvalJobs } from "../evaluation/evalService";
 import { OpenAIServer } from "./network";
 import { pruneDatabase } from "./utils";
