+++ worker/src/queues/blobStorageIntegrationQueue.ts
@@ -1,7 +1,7 @@
 import { Processor } from "bullmq";
 import { QueueJobs, logger } from "@langfuse/shared/src/server";
-import { handleBlobStorageIntegrationSchedule } from "../ee/integrations/blobstorage/handleBlobStorageIntegrationSchedule";
-import { handleBlobStorageIntegrationProjectJob } from "../ee/integrations/blobstorage/handleBlobStorageIntegrationProjectJob";
+import { handleBlobStorageIntegrationSchedule } from "../features/blobstorage/handleBlobStorageIntegrationSchedule";
+import { handleBlobStorageIntegrationProjectJob } from "../features/blobstorage/handleBlobStorageIntegrationProjectJob";
 
 export const blobStorageIntegrationProcessor: Processor = async (job) => {
   if (job.name === QueueJobs.BlobStorageIntegrationJob) {
